divide top level protocol packet implementations into submodules
